Swap field memory order Kernel|User Time.
- Get UserTime from the most significat side and Kernel from the lower side.
Weird, struct fields are usually laid in order

- Fix TdhContext.ParameterValue uint32 to unit64 to match the C interface size.

+++ etw/tdh_headers.go
@@ -18,7 +18,7 @@ typedef struct _TDH_CONTEXT {
 */
 
 type TdhContext struct {
-	ParameterValue uint32
+	ParameterValue uint64
 	ParameterType  TdhContextType
 	ParameterSize  uint32
 }
